Use CDN for faster image handling (#4)
Github's raw file storage is not a CDN and therefore is slow in many regions of the world. For example, it takes 2-4 seconds to fetch an image from Germany, Europe.
Therefore I'm experimenting with uploading and fetching the images from imagekit.io. They offer a generous free plan. If we find that performance is not satisfactory, we can consider other options like Google Cloud Storage.